Added to DeckofCards Class
