--- src/main/java/com/github/zipcodewilmington/casino/games/deckOfCards.java
@@ -1,22 +1,0 @@
-package com.github.zipcodewilmington;
-import java.util.ArrayList;
-
-public class DeckOfCards {
-    private ArrayList<Card> deck;
-
-
-
-    public DeckOfCards(ArrayList<Card> deck) {
-        this.deck = deck;
-    }
-
-
-
-    public DeckOfCards(){
-        List<String> suits =
-
-        List<String> faceNames =
-
-        deck = new ArrayList<Card>();
-    }
-}
